Fixes TypeError: document.querySelector(...) is null in jquery.main.js
The script is trying to find an element with class .close inside
another with class .toggleFrontEndModal which, from the source code,
it doesn't seem to exist (hence the error). There is one element with
class .toggleFrontEndModal that also has the class .close, which I'm
guessing is what this script is actually trying to hook into. This
commit fixes that.