Fixes PHP notice in page-settings.php
The code assumes that the variable savedRoleQuota is an array. This
won't be the case if the user hasn't previously saved role upload quotas
before (eg. fresh install) which will trigger some PHP notices due to
the code trying to access values in an array that hasn't been set yet.